Update to latest Snappy library.
The Snappy library structure was changed to simplify the import
path, code used to be rooted in /snappy/snappy, it is now rooted
in /snappy alongside the readme, etc.  Changing to use the latest
Snappy library to avoid conflicts with projects that rely on
libraries which use later versions of Snappy and the new
import path.

The only semi-substantative change that needed to be made was to
reflect the changed signature of Encode, which no longer returns
an error as encoding cannot fail.